+++ frontend/src/components/common/Text/MediaText.ts
