+++ frontend/src/pages/Profile.tsx
@@ -5,7 +5,7 @@ import Box from '../components/common/Box';
 import Flex from '../components/common/Flex';
 import Space from '../components/common/Space';
 import MediaSpace from '../components/common/Space/MediaSpace';
-import MediaText from '../components/common/Text/MediaQueryText';
+import MediaText from '../components/common/Text/MediaText';
 import MyInfo from '../components/MyInfo';
 import TopicCardContainerSkeleton from '../components/Skeletons/TopicListSkeleton';
 import { FULLSCREEN } from '../constants';
